Added TestThread in utils.py for future race-conditions testing in repositories, renamed integrated/template_module/factories.py to fakers.py, added empty test_commands.py for structural purposes, added TestThread to __all__
