+++ backend/tests/integrated/__init__.py
@@ -1,0 +1,3 @@
+from .utils import TestThread
+
+__all__ = ["TestThread"]
